requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
matplotlib>=3.7
opencv-python>=4.8
scikit-learn>=1.3
seaborn>=0.12
pandas>=2.0
wandb>=0.15
higher>=0.2
tqdm>=4.65
pyyaml>=6.0
tensorboard>=2.13
